view count fixed and script added

+++ boloindya/forum/user/models.py
@@ -76,7 +76,7 @@ class UserProfile(models.Model):
     share_count = models.PositiveIntegerField(null=True,blank=True,default=0)
     like_count = models.PositiveIntegerField(null=True,blank=True,default=0)
     vb_count = models.PositiveIntegerField(null=True,blank=True,default=0)
-    view_count = models.PositiveIntegerField(null=True,blank=True,default=0)
+    view_count = models.BigIntegerField(null=True,blank=True,default=0)
     bolo_score = models.PositiveIntegerField(null=True,blank=True,default=0)
     encashable_bolo_score = models.PositiveIntegerField(null=True,blank=True,default=0)
     is_geo_location = models.BooleanField(default=False)
